Adjust branch header parser
